pyproject.toml
[tool.poetry]
name = "iis-projekt"
version = "0.1.0"
description = ""
authors = ["jernej10 <jernej.ofic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
scikit-learn = "^1.5.0"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pytest = "^8.2.1"
yfinance = "^0.2.40"
great-expectations = "0.18.13"
evidently = "^0.4.25"
tensorflow = "^2.15.0.post1"
mlflow = "^2.13.0"
tensorflow-model-optimization = "^0.8.0"
tf2onnx = "^1.16.1"
dagshub = "^0.3.27"
tf-keras = "^2.16.0"
skl2onnx = "^1.16.0"
onnxruntime = "1.17.3"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
pymongo = "^4.7.2"
xgboost = "^2.0.3"
onnxmltools = "^1.12.0"

[tool.poetry.group.win-dev.dependencies]
tensorflow-intel = "^2.16.1"


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.26.1"
black = "^24.4.2"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
fetch-data = "python3 -m src.data.fetch_data"
process-data = "python3 -m src.data.process_data"
validate = "python3 -m src.data.validate"
data_drift = "python3 -m src.data.data_drift"
stability_tests = "python3 -m src.data.stability_tests"
ks_test = "python3 -m src.data.ks"
split_data = "python3 -m src.data.split_data"
train = "python3 -m src.models.train_model_test"
predict = "python3 -m src.models.predict_model_test"
